--- src/component/faq/FAQListTableComponent.jsx
@@ -1,30 +1,0 @@
-
-import {getFAQList, getFAQOne, updateFAQ} from "../../api/faqAPI";
-import CommonTableComponent from "@/common/CommonTableComponent.jsx";
-
-
-const column = [
-    "fno", "fcategory", "ftitle"
-];
-
-const tableHeader = [
-    "category", "title"
-]
-
-function FAQListTableComponent({  }) {
-
-
-    return (
-        <div>
-            <CommonTableComponent
-                tableHeader={tableHeader}
-                column={column}
-                listFn={getFAQList}
-                detailFn={getFAQOne}
-                updateFn={updateFAQ}
-            >
-            </CommonTableComponent>
-        </div>);
-}
-
-export default FAQListTableComponent;
